requirements.txt
altair==5.4.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==2.0.12
click==8.1.7
Flask==3.0.3
Flask-Cors==3.0.10
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib-metadata==6.11.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.4.2
numpy==1.26.4
packaging==23.2
pandas==2.2.2
Pillow==9.5.0
protobuf==4.25.4
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
streamlit==1.37.0
streamlit-option-menu==0.3.13
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.19
validators==0.33.0
watchdog==4.0.2
Werkzeug==3.0.3
zipp==3.20.0
